Code Review
Overall this code gets the job done and works pretty well.

I though the approach of using the totals to seperate the data was clever,
I would have assumed that each column would have been mapped to cost / GB usage based
based on the header.

The major issue I have with it is that it works for the current data set
but would break if any customers were added or removed.

Nice Job!